Correct deprecated and other issues with numpy 1.11.

+++ model/lib.py
@@ -26,8 +26,8 @@ def _cgrid_rho_vel(rho, dim, fill):
     if fill:
         rho = seapy.convolve_mask(rho, copy=True)
     shp = np.array(rho.shape)
-    fore = np.product([shp[i] for i in np.arange(0, dim)])
-    aft = np.product([shp[i] for i in np.arange(dim + 1, rho.ndim)])
+    fore = np.product([shp[i] for i in np.arange(0, dim)]).astype(int)
+    aft = np.product([shp[i] for i in np.arange(dim + 1, rho.ndim)]).astype(int)
     nfld = 0.5 * (rho.reshape([fore, shp[dim], aft])[:, 0:-1, :].filled(np.nan) +
                   rho.reshape([fore, shp[dim], aft])[:, 1:, :].filled(np.nan))
     shp[dim] = shp[dim] - 1
@@ -91,7 +91,7 @@ def u2rho(u, fill=False):
     shp = np.array(u.shape)
     nshp = shp.copy()
     nshp[-1] = nshp[-1] + 1
-    fore = np.product([shp[i] for i in np.arange(0, u.ndim - 1)])
+    fore = np.product([shp[i] for i in np.arange(0, u.ndim - 1)]).astype(int)
     nfld = np.ones([fore, nshp[-1]])
     nfld[:, 1:-1] = 0.5 * \
         (u.reshape([fore, shp[-1]])[:, 0:-1].filled(np.nan) +
@@ -122,7 +122,7 @@ def v2rho(v, fill=False):
     shp = np.array(v.shape)
     nshp = shp.copy()
     nshp[-2] = nshp[-2] + 1
-    fore = np.product([shp[i] for i in np.arange(0, v.ndim - 2)])
+    fore = np.product([shp[i] for i in np.arange(0, v.ndim - 2)]).astype(int)
     nfld = np.ones([fore, nshp[-2], nshp[-1]])
     nfld[:, 1:-1, :] = 0.5 * \
         (v.reshape([fore, shp[-2], shp[-1]])[:, 0:-1, :].filled(np.nan) +
